Move remaining app behavior to capability adapters
