+++ internal/app/router.go
@@ -56,8 +56,8 @@ func (s *Server) Routes() http.Handler {
 		r.Get("/admin/queries", s.protected(access.PermissionAuditRead, adminHTTP.Queries))
 		r.Get("/admin/queries/updates", s.protected(access.PermissionAuditRead, adminHTTP.QueryUpdates))
 		r.Post("/admin/queries/command", s.protected(access.PermissionAuditRead, adminHTTP.QueryCommand))
-		r.Post("/workspaces/{workspace}/access/upsert", s.protected(access.PermissionRBACManage, s.upsertWorkspaceAccess))
-		r.Post("/workspaces/{workspace}/access/remove", s.protected(access.PermissionRBACManage, s.removeWorkspaceAccess))
+		r.Post("/workspaces/{workspace}/access/upsert", s.protected(access.PermissionRBACManage, workspaceHTTP.AccessUpsert))
+		r.Post("/workspaces/{workspace}/access/remove", s.protected(access.PermissionRBACManage, workspaceHTTP.AccessRemove))
 		r.Get("/workspaces/{workspace}/permissions", s.protected(access.PermissionRBACManage, workspaceHTTP.Permissions))
 		r.Post("/workspaces/{workspace}/permissions", s.protected(access.PermissionRBACManage, workspaceHTTP.PermissionUpdate))
 		r.Post("/workspaces/{workspace}/permissions/remove", s.protected(access.PermissionRBACManage, workspaceHTTP.PermissionRemove))
