+++ internal/workspace/http/handler_test.go
@@ -1,4 +1,4 @@
-package app
+package http
 
 import (
 	"testing"
